Empezando los test

+++ Parcial/src/main/java/com/k1/Parcial/domain/service/servicesImpl/CustomerServiceImpl.java
@@ -3,6 +3,7 @@ package com.k1.Parcial.domain.service.servicesImpl;
 import com.k1.Parcial.application.request.Customer.CustomerPostDto;
 import com.k1.Parcial.application.request.Customer.CustomerUpdateDto;
 import com.k1.Parcial.domain.repository.CustomerRepository;
+import com.k1.Parcial.domain.service.ServiceException;
 import com.k1.Parcial.domain.service.serviceInterfaces.CustomerService;
 import com.k1.Parcial.domain.service.serviceInterfaces.EmployeService;
 import com.k1.Parcial.infrastructure.entity.Customer;
@@ -10,6 +11,7 @@ import com.k1.Parcial.infrastructure.entity.Employe;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -29,12 +31,18 @@ public class CustomerServiceImpl implements CustomerService {
     }
 
     @Override
-    public Optional<Customer> getById(Long id) {
-        return Optional.of(customerRepository.getById(id).orElseThrow());
+    public Optional<Customer> getById(Long id) throws ServiceException {
+        Optional<Customer> customer = customerRepository.getById(id);
+        if (customer.isPresent()) {
+            return customer;
+        } else {
+            throw new ServiceException("Customer not found");
+        }
     }
 
     @Override
     public void delete(Long id) {
+        if (Objects.isNull(id)) throw new RuntimeException("El id no puede ser nulo");
         customerRepository.delete(id);
     }
